Move product data to contract_compassion
