+++ contract_compassion/model/product.py
@@ -10,9 +10,28 @@
 ##############################################################################
 
 from openerp.osv import orm, fields
-from openerp.tools.translate import _
 
 
 class product(orm.Model):
     _inherit = 'product.product'
 
+    def _get_gmc_name(self, cr, uid, ids, field_name, arg, context=None):
+        res = dict()
+        gmc_names = {
+            'Birthday Gift': 'BirthdayGift',
+            'General Gift': 'GeneralChildGift',
+            'Family Gift': 'FamilyGift',
+            'Project Gift': 'ProjectGift',
+            'Graduation Gift': 'FinalOrGraduationGift'
+        }
+        for product in self.browse(cr, uid, ids, {'lang': 'en_US'}):
+            if product.product_tmpl_id.categ_id.name == 'Sponsor gifts':
+                res[product.id] = gmc_names[product.name]
+            else:
+                res[product.id] = "Undefined"
+        return res
+
+    _columns = {
+        'gp_fund_id': fields.integer("GP Fund id", size=4),
+        'gmc_name': fields.function(_get_gmc_name, type='char'),
+    }
