+++ sponsorship_compassion/model/__init__.py
@@ -13,7 +13,6 @@ from . import contract_group
 from . import contracts
 from . import invoice_line
 from . import res_partner
-from . import product
 from . import move_line
 from . import account_banking_mandate
 from . import project_compassion
